refactor(VCST-5159): extract MF wiring out of shared files into its own module
Make the Module Federation integration self-contained so the host's shared files
just import it, instead of inlining MF logic:

- vite.federation.ts (new, build scope): MF_SHARED + federatedHostPlugin() +
  federatedAlias(). vite.config.ts drops ~40 inlined lines → one import + a plugin
  spread + an alias spread. Lives outside client-app because it imports the
  @module-federation/vite dev dependency (build-time only).
- client-app/modules/federated/bootstrap.ts (new): startFederatedModules(), the
  app-runner entry (gated dynamic import). app-runner drops the inlined helper →
  one import + call + await.
- client-app/modules/federated/compare-versions.ts (new): the version-gate compare,
  local to the harness. Reverts the earlier export of core's compareVersions, so the
  harness now modifies ZERO core utilities.
- Shorten comments across the MF files.

Behaviour unchanged. Verified: validate:types (0), validate:dependencies (0),
APP_MF_HOST build (0, emits mf-manifest.json), MF files lint clean.
