+++ client-app/modules/federated/compare-versions.ts
@@ -1,0 +1,17 @@
+/**
+ * Dotted-numeric version compare for the federated version gate (VCST-5159, #2).
+ * Local to the harness so it needn't reach into core utilities. Returns <0 / 0 / >0.
+ */
+export function compareVersions(a: string, b: string): number {
+  const parse = (v: string) => v.split(".").map((p) => Number.parseInt(p, 10) || 0);
+  const pa = parse(a);
+  const pb = parse(b);
+  const length = Math.max(pa.length, pb.length);
+  for (let i = 0; i < length; i++) {
+    const diff = (pa[i] ?? 0) - (pb[i] ?? 0);
+    if (diff !== 0) {
+      return diff;
+    }
+  }
+  return 0;
+}
