+++ vite.config.ts
@@ -1,5 +1,4 @@
 import path from "path";
-import { federation } from "@module-federation/vite";
 import graphqlImport from "@rollup/plugin-graphql";
 import vue from "@vitejs/plugin-vue";
 import browserslistToEsbuild from "browserslist-to-esbuild";
@@ -7,6 +6,8 @@ import { visualizer } from "rollup-plugin-visualizer";
 import { defineConfig, loadEnv } from "vite";
 import { checker } from "vite-plugin-checker";
 import mkcert from "vite-plugin-mkcert";
+// Module Federation host config (VCST-5159) — split into vite.federation.ts.
+import { federatedHostPlugin, federatedAlias } from "./vite.federation.js";
 import type { ProxyOptions, UserConfig, PluginOption } from "vite";
 
 const graphql = graphqlImport.default ?? graphqlImport;
@@ -14,35 +15,6 @@ const graphql = graphqlImport.default ?? graphqlImport;
 // Libraries imported only via dynamic import()/defineAsyncComponent; kept out of the eager
 // `vendor` chunk so they stay in their own lazy chunks.
 const DEFERRED_LIBS = ["skyflow-js", "barcode-detector", "marked", "nouislider"];
-
-/**
- * Module Federation shared singletons (VCST-5159).
- *
- * INTENTIONALLY MINIMAL (#6): a package belongs here only if a second instance would
- * break correctness. Sharing pins the version host-wide — plugins bind to the host's
- * copy — so anything that does NOT strictly require a single instance is left OUT and
- * bundled per-plugin instead, preserving independent versioning. Each entry's reason:
- *   - vue / vue-router / vue-i18n : one framework instance (reactivity, the router the
- *     app navigates, the i18n messages are merged into) — two copies = broken inject.
- *   - @vueuse/core               : `createGlobalState` backs the shared state /
- *     extension registry; a second copy = a second, empty global state.
- *   - @apollo/client + @vue/apollo-composable + graphql : one client/cache/injection,
- *     and Apollo requires a single `graphql` (DocumentNodes from another copy throw).
- *   - @vc-frontend/core          : the host provides the live facade instance.
- *
- * `requiredVersion: "*"` keeps MF itself from blocking; real compatibility is enforced
- * explicitly by the loader's version gate (see client-app/modules/federated, #2).
- */
-const MF_SHARED = {
-  vue: { singleton: true, requiredVersion: "*" },
-  "vue-router": { singleton: true, requiredVersion: "*" },
-  "vue-i18n": { singleton: true, requiredVersion: "*" },
-  "@vueuse/core": { singleton: true, requiredVersion: "*" },
-  "@apollo/client": { singleton: true, requiredVersion: "*" },
-  "@vue/apollo-composable": { singleton: true, requiredVersion: "*" },
-  graphql: { singleton: true, requiredVersion: "*" },
-  "@vc-frontend/core": { singleton: true, requiredVersion: "*", version: "2.53.0" },
-} as const;
 
 function getProxy(target: ProxyOptions["target"], options: Omit<ProxyOptions, "target"> = {}): ProxyOptions {
   const dontTrustSelfSignedCertificate = false;
@@ -78,24 +50,8 @@ export default defineConfig(({ command, mode }): UserConfig => {
           })
         : undefined,
       vue(),
-      // Module Federation host (VCST-5159). Declares the shared singletons that
-      // plugins bind to and emits mf-manifest.json. Remotes are registered at
-      // runtime by the loader (client-app/modules/federated) from APP_MF_REMOTES,
-      // so none are hard-coded here. Gated behind APP_MF_HOST so the default
-      // dev/build is unchanged unless a host opts into acting as an MF host.
-      process.env.APP_MF_HOST
-        ? (federation({
-            name: "host",
-            filename: "remoteEntry.js",
-            manifest: true,
-            // Off: the dts plugin would emit .d.ts across the shared facade's whole
-            // source graph. Types are produced by the dedicated publish-from-source
-            // build instead (yarn build:core-types → core-api/dist/index.d.ts).
-            dts: false,
-            shareStrategy: "loaded-first",
-            shared: MF_SHARED,
-          }) as PluginOption)
-        : undefined,
+      // Module Federation host — empty unless APP_MF_HOST is set (see vite.federation.ts).
+      ...federatedHostPlugin(process.env.APP_MF_HOST),
       graphql() as PluginOption,
       isServe
         ? checker({
@@ -119,8 +75,8 @@ export default defineConfig(({ command, mode }): UserConfig => {
     resolve: {
       alias: {
         "@": path.resolve(__dirname, "client-app"),
-        // Host is the provider of the shared facade — it resolves to real source.
-        "@vc-frontend/core": path.resolve(__dirname, "client-app/core-api"),
+        // Host provides the shared facade — resolves @vc-frontend/core to real source.
+        ...federatedAlias(__dirname),
       },
     },
     define: {
